Refactor2022

+++ lib/kamishibai/functions.rb
@@ -143,18 +143,29 @@ def cbz_pages?( zfile )
 	return i
 end
 
+# allow natural sort for filename
+class String
+	def naturalized
+		scan(/[^\d\.]+|[\d\.]+/).collect { |f| f.match(/\d+(\.\d+)?/) ? f.to_f : f }
+	end
+end
 
 # cbz file accessor, give file name and page and you shall receive
 def open_cbz( zfile, page = 1, options = {} )
+	if !FileTest.exists?(zfile)
+		puts "error: cbz file not found #{zfile}"
+		return nil
+	end
+
 	objs = []
-	# begin
+	#begin
 		Zip::File.open( zfile ) { |x|
 			x.each { |zobj|
 				if zobj.ftype == :file and File.basename(zobj.name)[0] != '.' and File.basename( zobj.name ) =~ /\.(jpg|jpeg|png|gif)$/i
 					objs << zobj
 				end
 			}
-			objs.sort!
+			objs.sort_by! { |zobj| zobj.name.to_s.naturalized }
 
 			if objs.length == 0
 				puts "error: no image detected. #{zfile}"
